# src/docs/swagger.yaml
basePath: /
definitions:
  errors.ErrorDetails:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        description: ESTE NO IRIA MAS
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  errors.ErrorDetailsWithParams:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
      validation_errors:
        items:
          $ref: '#/definitions/errors.ErrorParam'
        type: array
    type: object
  errors.ErrorParam:
    properties:
      field:
        type: string
      reason:
        type: string
    type: object
...
  models.Book:
    properties:
      amount_of_pages:
        type: integer
      author:
        type: string
      author_name:
        type: string
      avg_rating:
        type: number
      description:
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        type: string
      language:
        type: string
      publication_date:
        type: string
      title:
        type: string
      total_ratings:
        type: integer
    required:
    - amount_of_pages
    - author
    - author_name
    - description
    - genres
    - id
    - language
    - publication_date
    - title
    type: object
...
  models.BookResponseWithReview:
    properties:
      book:
        $ref: '#/definitions/models.BookResponse'
      review:
        $ref: '#/definitions/models.Review'
      status:
        type: string
    type: object
...
  models.NewBookRequest:
    properties:
      amount_of_pages:
        type: integer
      description:
        type: string
      genres:
        items:
          type: string
        type: array
      language:
        type: string
      picture:
        items:
          type: integer
        type: array
      publication_date:
        type: string
      title:
        type: string
    required:
    - amount_of_pages
    - description
    - genres
    - language
    - publication_date
    - title
    type: object
  models.NewRatingRequest:
    properties:
      rating:
        type: integer
    required:
    - rating
    type: object
  models.NewReviewRequest:
    properties:
      rating:
        type: integer
      review:
        type: string
    required:
    - rating
    type: object
...
  models.ReviewOfBook:
    properties:
      publication_date:
        type: string
      rating:
        type: integer
      review:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a  server for Swagger with Gin.
  title: BetterReads API
  version: "1.0"
paths:
  /books:
    post:
      consumes:
      - multipart/form-data
      description: publishes a book, the book data should follow the models.NewBookRequest
        in JSON
      parameters:
      - description: Book Data
        in: formData
        name: data
        required: true
        type: string
      - description: Book Picture
        in: formData
        name: file
        required: true
        type: file
      - description: Don't need to send this in json, this param is only here to reference
          NewBookRequest, DONT SEND PICTURE in JSON
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.NewBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetailsWithParams'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: publish a book
      tags:
      - books
  /books/{id}/info:
    get:
      description: Get book id, note that its a UUID
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BookResponseWithReview'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Get book by id
      tags:
      - books
  /books/{id}/picture:
    get:
      description: Get book id, note that its a UUID
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Get book picture by id
      tags:
      - books
  /books/{id}/rating:
    post:
      consumes:
      - application/json
      description: Rate a book
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      - description: Rating Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.NewRatingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetailsWithParams'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Rate a book
      tags:
      - books
    put:
      description: Update rating of a book
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      - description: Rating Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.NewRatingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Update rating of a book
      tags:
      - books
  /books/{id}/review:
    delete:
      description: Delete review of a book that belongs to the user
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Delete review of a book
      tags:
      - books
    get:
      description: Get reviews of a book
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ReviewOfBook'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetailsWithParams'
      summary: Gets reviews of a book
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Add review to a book
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      - description: Review Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.NewReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetailsWithParams'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Add review to a book
      tags:
      - books
    put:
      description: Edit review of a book that belongs to the user
      parameters:
      - description: Book Id
        in: path
        name: id
        required: true
        type: string
      - description: Review Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.NewReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NewReviewRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetailsWithParams'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Edit review of a book
      tags:
      - books
  /books/author/{id}:
    get:
      description: Get the books of an author, if no books found returns an empty
        array
      parameters:
      - description: Author Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BookResponseWithReview'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Get books of an auther
      tags:
      - books
  /books/genres:
    get:
      description: Get all genres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Get all genres
      tags:
      - books
  /books/info:
    get:
      consumes:
      - application/json
      description: Get all books
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BookResponseWithReview'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Get all books
      tags:
      - books
  /books/info/search:
    get:
      description: Get books by name, filters by genre and also sorts, if no books
        found returns an empty array
      parameters:
      - description: Book Name
        in: query
        name: name
        required: true
        type: string
      - description: Book Genre
        in: query
        name: genre
        type: string
      - description: Sort by publication_date, total_ratings, avg_rating
        in: query
        name: sort
        type: string
      - description: Sort direction asc or desc
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BookResponseWithReview'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Get books by name
      tags:
      - books
...
swagger: "2.0"